==> marine/bivalve/bivalve_farms/STEP2_bivalve_farm_locations.Rmd <==
---
title: "Location of Bivalve Farms"
author: "Gage Clawson (NCEAS, UCSB, OHI)"
date: "12/14/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

This script standardizes location of bivalve farms to be lat/long point data.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(raster)
library(tidyverse)
library(stringr)
library(rgdal)
library(PBSmapping)
library(maptools)
library(sf)
library(lwgeom)
library(data.table)
library(here)
library(mapview)
library(janitor)
library(rnaturalearth)
library(spatialEco)
library(countrycode)
# library(remotes)
# install_github("yonghah/esri2sf")
library("esri2sf")


path <- "/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations"

fao_bivalve_spp <- read_csv("marine/bivalve/bivalve_farms/data/fao_mariculture_bivalve_taxa_production.csv") %>%
  dplyr::filter(year == 2017)

fao_bivalve <- read_csv("marine/bivalve/bivalve_farms/data/fao_mariculture_bivalve.csv") %>%
  dplyr::filter(year == 2017)

unique(fao_bivalve$country)
#  [1] "Australia"                "Brazil"                   "Bulgaria"                 "Cambodia"                
#  [5] "Canada"                   "Chile"                    "China"                    "China, Hong Kong SAR"    
#  [9] "Croatia"                  "Cuba"                     "Denmark"                  "France"                  
# [13] "French Polynesia"         "Germany"                  "Greece"                   "Guernsey"                
# [17] "India"                    "Indonesia"                "Ireland"                  "Italy"                   
# [21] "Japan"                    "Jersey"                   "Korea, Dem. People's Rep" "Korea, Republic of"      
# [25] "Malaysia"                 "Mexico"                   "Netherlands"              "New Zealand"             
# [29] "Norway"                   "Peru"                     "Philippines"              "Portugal"                
# [33] "Russian Federation"       "Slovenia"                 "South Africa"             "Spain"                   
# [37] "Sweden"                   "Taiwan Province of China" "Thailand"                 "United Kingdom"          
# [41] "United States of America" "Viet Nam"       

```


# Raster

Standard raster used for this analysis 

```{r}

source(here::here("_spatial/template_raster.R"))

```

# Global regions

```{r}

gadm <- raster("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm36mol_pyfill_10km.tif")

gadm_latlon <- projectRaster(gadm, food_raster)

gadm_codes <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm_reg_codes.csv") %>% 
  mutate(reg_id = as.numeric(reg_id))

```

```{r}

gadm_df <- tabularaster::as_tibble(gadm_latlon, value = TRUE, cells = TRUE) %>% 
  mutate(cellvalue = round(cellvalue, 0))

gadm_reg <- dplyr::left_join(gadm_df, gadm_codes, by = c("cellvalue" = "reg_id"))

```

Now we will map the remaining countries with bivalve farms. Workflow will be as follows: 
 - Try to obtain exact locations of farms from published source or dataset (such as FAO or a research paper)
 - If there are not exact locations or comprehensive locations, find the number of farms. 
 - If there is no information at all, note it. We will gapfill later. 
 
For each point, we will add a column for "data quality". It will be scored as very good (5), good (4), fair (3), bad (2), and very bad (1). 

5: Comprehensive number of exact farms 

4: Subset of exact locations, need to add a specific number of farms to existing locations. 
  - country|species|farm locations
  - country|species|additional points

3: A specific number of farms are known, but no locations 
 - country|species|additional points
 
2.5: A specific number of farms that are not comprehensive, so we will have to gapfill on top of those, if necessary. 
 - country, species, n_existing_points

2: Exact locations are a subset of total farms. But the number of total farms are unknown. Need to be gapfilled by tonnage. 

1: No information on number of farms at all. Need to be gapfilled by tonnage. 
 - country|species|# additional farms (this will be NA)

## Read in ASC data
https://gis.asc-aqua.org/arcgis_app/
Maps from ASC

```{r}
ASC_data <- st_read(file.path("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/ASC_data/ASC_all_sf.shp"))

ASC_bivalve <- ASC_data %>%
  dplyr::filter(aq_grop == "bivalves")

unique(ASC_bivalve$crt_cnt)
#  [1] "China"                                               
#  [2] "New Zealand"                                         
#  [3] "South Africa"                                        
#  [4] "Australia"                                           
#  [5] "Korea, Republic of"                                  
#  [6] "Japan"                                               
#  [7] "Peru"                                                
#  [8] "Mexico"                                              
#  [9] "France"                                              
# [10] "Chile"                                               
# [11] "United Kingdom of Great Britain and Northern Ireland"
# [12] "Viet Nam"                                            
# [13] "Portugal"  


## what is in ASC and not in FAO
setdiff(ASC_bivalve$crt_cnt, fao_bivalve$country)
#[1] "United Kingdom of Great Britain and Northern Ireland"

## what is in FAO and not ASC -- this is what we need to find alternative data sources for (29 countries)
setdiff(fao_bivalve$country, ASC_bivalve$crt_cnt)
#  [1] "Brazil"                   "Bulgaria"                 "Cambodia"                 "Canada"                  
#  [5] "China, Hong Kong SAR"     "Croatia"                  "Cuba"                     "Denmark"                 
#  [9] "French Polynesia"         "Germany"                  "Greece"                   "Guernsey"                
# [13] "India"                    "Indonesia"                "Ireland"                  "Italy"                   
# [17] "Jersey"                   "Korea, Dem. People's Rep" "Malaysia"                 "Netherlands"             
# [21] "Norway"                   "Philippines"              "Russian Federation"       "Slovenia"                
# [25] "Spain"                    "Sweden"                   "Taiwan Province of China" "Thailand"                
# "United States of America"


mapview(ASC_bivalve)
```


# Map data from ASC 

Data comes from https://gis.asc-aqua.org/arcgis_app/ 

We filter out france and great britain, because those farms are already accounted for in the EMODnet database (which are comprehensive datasets taken from lease data). 

Source: 
Aquaculture Stewardship Council Data Mapping Study, n.d. Aquaculture Stewardship Council. [Accessed November 11, 2020] 
https://gis.asc-aqua.org/arcgis_app/


```{r}

# get iso3c codes
iso <-   data.frame(iso3c = countrycode(ASC_bivalve$crt_cnt, origin = 'country.name', destination = 'iso3c'))


# read in shp
farms_all <- ASC_bivalve %>%
  cbind(iso) %>%
  mutate(iso3c = ifelse(str_detect(comp_nm, "Jersey"), "JEY", iso3c)) %>% ## fix Jersey
  filter(!(iso3c %in% c("GBR", "FRA"))) ## filter out france and GBR... because they are in EMODnet

mapview(farms_all$geometry)

unique(farms_all$iso3c)

all_pts <- st_centroid(farms_all)

all_pts_latlon <- st_transform(all_pts, crs = "+init=epsg:4326") %>% 
  dplyr::select(geometry, iso3c)

farms_all_fin <- data.frame(do.call(rbind, st_geometry(all_pts_latlon))) 
names(farms_all_fin) <- c("lon", "lat")
farms_all_fin <- farms_all_fin %>%
  cbind(iso3c = all_pts_latlon$iso3c)

## save farm location
write.csv(farms_all_fin, here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv"), row.names = FALSE)

```


Now we need to map for each country, and assign a flag for gapfilling which will be completed later. I will start with all of the countries list in the ASC data. 

# China 

Current data is from ASC and https://zenodo.org/record/3881612, only 11 + 4737 = 4748 farms, so probably not comprehensive, just a drop in the bucket compared to the production of China. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

Fu, Y., Deng, J., Wang, H., Comber, A., Yang, W., Wu, W., You, S., Lin, Y., Wang, K., 2021. A new satellite-derived dataset for marine aquaculture areas in China’s coastal region. Earth Syst. Sci. Data 13, 1829–1842. https://doi.org/10.5194/essd-13-1829-2021

Fu et al., 2021

```{r}
farms_chn <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "CHN") %>%
  mutate(data_quality = 2)

farms_chn_2 <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/China/split_spp/farms_biv_chn.csv") %>%
  dplyr::select(-id) %>%
  mutate(data_quality = 2)

farms_chn_fin <- rbind(farms_chn, farms_chn_2)

write_csv(farms_chn_fin, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_chn.csv"))


```

# New Zealand

We currently know the exact location of 2 bivalve farms in New Zealand, however there is much more production that just 2 farms in New Zealand. Although outdated, the FAO report tells us there are 910 total abaolne, mussel, and oyster farms in New Zealand. We will treat this as truth. 

 -  Need to add 908 points to existing 2 (a final total of 910 points)

 - http://www.fao.org/fishery/countrysector/naso_newzealand/en

  - “In 2004 there were 645 New Zealand mussel farms occupying a total of 4 747 hectares of water space and producing over 85 000 tonnes of mussels (NZAC, 2005). The majority of these farms were in the Marlborough Sounds at the northern end of the South Island. A smaller proportion of the mussel farms are located in the Coromandel and Hauraki Gulf in the north of the North Island, and a smaller number again at Stewart Island. More recently, permission has been given to develop farming in some larger offshore areas in the Bay of Plenty and Hawkes Bay off the eastern coast of the North Island.”

  - “In 2004 there were 230 oyster farms covering 750 hectares of marine waters mostly in a series of sheltered harbours on the north-eastern coast of the North Island (NZAC, 2005). These farms produced around 3 300 tonnes of oysters.”

  - “A further 35 abalone farms operated over 20 hectares of marine water space producing only 2 tonnes of product in 2004 (NZAC, 2005). Most of these farms were located in the Marlborough Sounds or on Banks Peninsula on the eastern coast of the South Island.”


```{r}
farms_nzl <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "NZL") %>%
  mutate(data_quality = 4)

write.csv(farms_nzl, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_nzl.csv"), row.names = FALSE)


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
nz_additional <- data.frame(iso3c = "NZL", type = "bivalve", num_additional_farms = 908, data_quality = 4, total_farms = 910)
write.csv(nz_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_nz.csv"), row.names = FALSE)
```

# South Africa 

We currently know the exact location of 2 bivalve farms in South Africa, however there is much more production that just 2 farms. However, governemnt websites and research papers say that there are 4 mussel farms and 21 abalone farms. We will treat this as truth. 

 - Need to add 23 points to existing 2 (a final total of 25 points)
 - https://southafrica.co.za/mussel-production.html
   - “According to The Department of Agriculture, Forestry and Fisheries (DAFF), mussel farming in South Africa is conducted entirely in Saldanha Bay, Western Cape, where four mussel farms are operational.”
 - https://core.ac.uk/download/pdf/62634432.pdf
  - We can tell from Figure 1B that there are 21 abalone farms in South Africa


```{r}
farms_zaf <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "ZAF") %>%
  mutate(data_quality = 4)

write.csv(farms_zaf, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_zaf.csv"), row.names = FALSE)


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
zaf_additional <- data.frame(iso3c = "ZAF", type = "bivalve", num_additional_farms = 23, data_quality = 4, total_farms = 25)
write.csv(zaf_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_zaf.csv"), row.names = FALSE)
```

# Australia  

We currently know the exact location of 1 ASC bivalve farm, and 23 oyster farms in Australia, however there is much more production that just 24 farms. Aquaculture websites say there are 20 abalone farms, 23 oyster farms, and 25 mussel farms in Australia 

Need to add 45 points to existing 24 points (for a total of 69 points)

 - Abalone
  - https://www.oceanwatch.org.au/aquaculture-species/
  - “There are around 20 farms that produce about 725 tonnes of abalone every year, with production volume increasing by about 25% every year.”
  - 725 tonnes of abalone matches really well to what is reported for abalone in Australia by the FAO. 

 - Oysters
  - This states that nearly all oyster license holders are a part of this association:   
  - https://www.aquaculturealliance.org/advocate/oyster-farming-in-south-australia/
  - We will use this as truth: https://oysterssa.com.au/oyster-growers/
  - We will need to place an additional 23 oyster farms in Australia 

 - Mussels
  - https://www.oceanwatch.org.au/aquaculture-species/
  - “Concentrated in the south of country, there are about 25 mussel farms spread across SA, Tasmania, Victoria, Tasmania and the south coasts of both NSW and WA.”

  - We will place a total of 69 bivalve points in Australia. 


```{r}
farms_AUS <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "AUS") %>%
  mutate(data_quality = 4) ## this is an abalone farm

## now we need to add the oyster farms from this map: https://oysterssa.com.au/oyster-growers/
farms_AUS_2 <- data.frame(lat_lon = c("-32.377148325671286, 133.93965198614558", "-34.91844207220454, 137.7783987555795", "-35.77943962962783, 137.77191642975748", "-33.693162657803185, 136.924417900667", "-32.37607854727286, 133.94039259834852", "-32.373416358293625, 133.9357013204693", "-32.80283080755049, 134.1850786971314", "-34.62455022926624, 135.46366166621291", "-33.67979293746246, 136.91678920994693", "-34.551319592404916, 135.3981400818773", "-34.61868048533167, 135.4628853832039", "-32.3774179110106, 133.94079648728368", "-33.71698462803702, 136.90259542838538", "-34.90762372887531, 137.7836548109512", "-34.623471540228394, 135.46192328133426", "-33.67594103960825, 136.92673532055025", "-32.37775541730595, 133.94005795577584", "-34.52175570620221, 135.38834856713308", "-33.677446237574934, 136.92623952576452", "-32.37683026324807, 133.94078430357064", "-34.625125897259146, 135.46296226954271", "-34.61986531053386, 135.46279518888247", "-32.3779085349619, 133.9409211712551")) %>%
  separate(col = lat_lon, sep = ", ", into = c("lat", "lon")) %>% 
  mutate(iso3c = "AUS") %>%
  mutate(data_quality = 4)

all_farms_AUS <- rbind(farms_AUS, farms_AUS_2)


write.csv(all_farms_AUS, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_AUS.csv"), row.names = FALSE)


all_farms_sf <- st_as_sf(all_farms_AUS, coords = c("lon", "lat"),
                         crs = crs(gadm_latlon), agr = "constant")

mapview(all_farms_sf)

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
AUS_additional <- data.frame(iso3c = "AUS", type = "bivalve", num_additional_farms = 45, data_quality = 4, total_farms = 69)
write.csv(AUS_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_aus.csv"), row.names = FALSE)
```

# South Korea 
We currently know the exact location of 27 bivalve farms in South Korea. However, FAO says there are 5245 shell fish leases in South Korea. WE will place an additional 5218 farms. 


Need to add 5218 points to existing 27 (a final total of 5245 points)
 - FAO states there are 5,245 shellfish leases in operation:     
  - http://www.fao.org/fishery/countrysector/naso_korea/en


```{r}
farms_kor <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "KOR") %>%
  mutate(data_quality = 4)

write_csv(farms_kor, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_kor.csv"))


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
kor_additional <- data.frame(iso3c = "KOR", type = "bivalve", num_additional_farms = 5218, data_quality = 4, total_farms = 5245)
write.csv(kor_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_kor.csv"), row.names = FALSE)
```

# Japan

Current data is from ASC, only 55 farms, so definitely not comprehensive, just a drop in the bucket compared to the production of Japan. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_jpn <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "JPN") %>%
  mutate(data_quality = 2)

write_csv(farms_jpn, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_jpn.csv"))

```

# Peru 

We currently know the exact location of 25 bivalve farms in Peru. However, FAO says there are 52 bivalve farms in Peru. We will place an additional 27 farms (for a total of 52 farms). 

 - Need to add 27 points to existing 25 (a final total of 52 points)
   - http://www.fao.org/fishery/countrysector/naso_peru/en
   - “It is estimated that Peru has some 14 164 ha apt for the development of marine cultures; of which 86 percent correspond to the region of Ancash (7 228.14ha) and to Ica (5 017.97 ha). Nevertheless, as of December 2003, leased and authorized areas amount to only 8 858 ha. Of this total, 3 033 ha are dedicated to the cultivation of the fan shell.
   - At present, of the 44 referred enterprises, 26 holding 1 662.8 ha are located in the Ancash region; followed by the Ica region, where 18 enterprises occupy 136.8 ha; and the Piura Region has 8 enterprises that develop their activities in 787,6 ha. The majority are large scale operations, with the exception of those located in the Ica region which are relatively small.”


```{r}
farms_per <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "PER") %>%
  mutate(data_quality = 4)

write_csv(farms_per, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_per.csv"))


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
per_additional <- data.frame(iso3c = "PER", type = "bivalve", num_additional_farms = 27, data_quality = 4, total_farms = 52)
write.csv(per_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_per.csv"), row.names = FALSE)
```


# Mexico
Current data is from ASC, only 1 farm, so definitely not comprehensive, just a drop in the bucket compared to the production of Mexico. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_mex <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "MEX") %>%
  mutate(data_quality = 2)

write_csv(farms_mex, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_mex.csv"))

```


# Jersey
Current data is from ASC, only 5 farm. However, this is comprehensive. When comparing to governmental lease data, there are 5 leases in Jersey. 

```{r}
farms_jey <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "JEY") %>%
  mutate(data_quality = 5)

write_csv(farms_jey, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_jey.csv"))

```

# Viet Nam

Current data is from ASC, only 1 farm, so definitely not comprehensive, just a drop in the bucket compared to the production of Viet Nam. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_vnm <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "VNM") %>%
  mutate(data_quality = 2)

write_csv(farms_vnm, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_vnm.csv"))

```

# Portugal


We currently know the exact location of 1 bivalve farm in Portugal, however there is much more production that just 1 farm. However, Portuguese government data says there are 30 bivalve farms in Portugal We will place an additional 29 farms (for a total of 30 farms). 


Need to add 29 points to existing 1 (a final total of 30 points)

Current data is ASC data which is only a subset of total farms. 

 - https://www.ine.pt/xportal/xmain?xpid=INE&xpgid=ine_publicacoes&PUBLICACOESpub_boui=435690295&PUBLICACOESmodo=2&fbclid=IwAR0-_SHYtw2XGUAXdx71RGer284Vk_u40HYKXyJzeFbpdnsHydKgAfomTvU
  - National Institute of Statistics for Portugal 
  - There is a csv table, "q5_1.csv", which has the number of farms listed by type of system. I will assume that floating marine structures are for bivalves, since [this website](https://rea.apambiente.pt/content/aquaculture-production?language=en) says that the tanks are for **fish production**, and floating structures are for **bivalve molluscs**.
  
  - Located here "/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/Portugal/EPesca_2019.xlsx"

 - According to this table, there are 30 total marine floating structures  

```{r}
farms_prt <- read.csv(here("marine/bivalve/bivalve_farms/data/ASC_bivalve_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "PRT") %>%
  mutate(data_quality = 4)

write_csv(farms_prt, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_prt.csv"))


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
prt_additional <- data.frame(iso3c = "PRT", type = "bivalve", num_additional_farms = 29, data_quality = 4, total_farms = 30)
write.csv(prt_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_prt.csv"), row.names = FALSE)
```


# Chile
This data is comprehensive. No gapfilling needed. 

Data for Chile comes from https://mapas.subpesca.cl/ideviewer/?locale=en-gb. 

There is also a list of concessions here: http://www.subpesca.cl/portal/619/w3-article-92935.html

I was in contact with Valeria Tamayo at UCSB who helped me with some spanish translations concerning this data.

Source:
Undersecretariat for Fisheries and Aquaculture, 2020. [Accessed Feb 2020] https://mapas.subpesca.cl/ideviewer/?locale=en-gb

```{r}

# Column meaning and names
#0 LINK_DOC - URL of the SUBPESCA resolution approving the aquaculture or resolution technical project that cartographically regulates the aquaculture concession
#1 CARTA - Name of the geographic reference letter
#2 DATUM - Reference datum name
#3 HUSO - Spindle Name
#4 COMUNA - Name of the commune
#5 N_PERT - Pert number, code of the application for admission to process
#6 N_CODIGOCE - Farming center code according to the national aquaculture registry
#7 N_RESOLSSP - Resolution number of the undersecretary of fisheries and aquaculture that apprehends the aquaculture technical project
#8 F_RESOLSSP - Date of issuance of the resolution of the undersecretary of fisheries and aquaculture approving the aquaculture technical project
#9 N_RESOLSSF - Resolution number of the undersecretary for the armed forces granted by the aquaculture concession
#10 F_RESOLSSF - Date of issuance of the resolution of the undersecretary for the armed forces granted by the aquaculture concession
#11 TOPONIMIO - Geographical location of the aquaculture concession
#12 ESPECIES - Name of crop species
#13 AGRUPCONCE - Identifier of the salmonid concession group or sanitary neighborhood
#14 SUPERFICIE - Surface in hectares
#15 TITULAR - Name of the holder of the aquaculture concession
#16 COORDENADS - List of geographical coordinates of the aquaculture concession corresponding to each of the vertices of the polygon
#17 REGION - Region name
#18 GRUPO_ESP - Group of crop species
#19 TIPO_PORCI - Type of space where the aquaculture concession is located
#20 ESTADO - Processing status of the aquaculture concession
#21 ESTADO_SOL - Status of the aquaculture application process


chile_dat <- st_read(file.path(path, "Chile"), layer = "Concesiones_de_Acuicultura")

```

We filter to only those that produce salmon, but note that this is ALL leases, not neccesarily only active sites. 
```{r}
unique(chile_dat$REP_SUB_13)


#Filter to only bivalve farming (mussel, abalone, oysters, clams, and scallops)
chile <- chile_dat %>% 
  filter(grepl("CHORO|ABALON|OSTRA|CHOLGA|CHORITO|OSTION|ALMEJA|MITILIDOS", REP_SUB_13)) %>%
  filter(!grepl("SALMON", REP_SUB_13)) 

# convert polygons to points
chile_pts <- st_centroid(chile)
chile_pts <- st_transform(chile_pts, food_crs)

# save as a dataframe
farms_chile <- data.frame(do.call(rbind, st_geometry(chile_pts))) 
names(farms_chile) <- c("lon", "lat")
farms_chile <- farms_chile %>%
  mutate(iso3c="CHL") %>%
  mutate(data_quality = 5) 

write_csv(farms_chile, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_chl.csv"))

```


# USA 

Based on the datasets below, we have an existing 1010 bivalve farms. We need to add 193 points to existing 1010 (a final total of 1203 points)

Our current information is taken from many different state level sources, but we are definitely still missing some, as California is not included... we have farm/lease locations from 8 different states:

 - Alaska
 - Alabama
 - Connecticut
 - Maine
 - North Carolina
 - New Hampshire
 - Rhode Island
 - Washington

 - However, per FAO 2005, there were 1203 saltwater shellfish farms in the USA.. unclear if this is in marine environments, but it is the best we have. 
  - http://www.fao.org/fishery/countrysector/naso_usa/en


## Maine


 - Viewer: https://www.maine.gov/dmr/aquaculture/leases/aquaculturemap.html
 - Download: https://dmr-maine.opendata.arcgis.com/datasets/maine::mainedmr-aquaculture-aq-leases/about
 - I think we will want to use regular leases only.
 - Limited purpose leases are for scientific or smaller/shorter term mariculture projects, that last up to 3 years.
 - I think this can be considered "comprehensive". 
 - Update to date as of 2019.
 - Has species types specified. 
 
 
Locations of aquaculture leases in Maine, 2019. . Maine Department of Marine Resources. https://dmr-maine.opendata.arcgis.com/datasets/maine::mainedmr-aquaculture-aq-leases/about


```{r}
maine_aq_lease <- st_read("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/USA/Maine/MaineDMR_Aquaculture_AQ_Leases/MaineDMR_Aquaculture_AQ_Leases.shp")
  

# mapview(maine_aq_lease)

maine_aq_lease_active <- maine_aq_lease %>%
  dplyr::filter(STATUS == "A") %>% # 149 leases
  dplyr::filter(str_detect(PRIMARYSP, "Oyster|Mussel|Clam|mussels|Scallop")) %>%
  dplyr::filter(!str_detect(PRIMARYSP, "Salmon"))

unique(maine_aq_lease_active$PRIMARYSP)
  

us_me_bivalve_mol <- st_transform(maine_aq_lease_active, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_me_bivalve_pts <- st_centroid(us_me_bivalve_mol)
# check
plot(st_geometry(us_me_bivalve_mol))
plot(st_geometry(st_centroid(us_me_bivalve_mol)), add=TRUE)

us_me_bivalve_pts_latlon <- st_transform(us_me_bivalve_pts, crs = "+init=epsg:4326") 

farms_us_me <- data.frame(do.call(rbind, st_geometry(us_me_bivalve_pts_latlon))) # get lat lon
names(farms_us_me) <- c("lon", "lat") 
farms_us_me <- farms_us_me %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "Maine") 

mapview(us_me_bivalve_pts_latlon)
```



## Rhode Island

 - Lease map here: https://ridemgis.maps.arcgis.com/apps/webappviewer/index.html?id=8beb98d758f14265a84d69758d96742f
 - Current as of August 2021
 - Shellfish and kelp
 
RIDEM Marine Fisheries Maps Approved Aquaculture Areas in Rhode Island, 2021. . Rhode Island Department of Environmental Management. https://ridemgis.maps.arcgis.com/apps/webappviewer/index.html?id=8beb98d758f14265a84d69758d96742f

 
```{r}
rhode_island <- esri2sf("https://risegis.ri.gov/hosting/rest/services/RIDEM/Marine_Fisheries_Map_Layers/MapServer/0")
mapview(rhode_island)

ri_active_biv <- rhode_island %>%
  filter(GPS_Source %in% c("Approved", "Expanded", "PD App")) %>%
  filter(Oysters == "Yes" | Clams == "Yes" | Scallops == "Yes")

us_ri_bivalve_mol <- st_transform(ri_active_biv, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_ri_bivalve_pts <- st_centroid(us_ri_bivalve_mol)
# check
plot(st_geometry(us_ri_bivalve_mol))
plot(st_geometry(st_centroid(us_ri_bivalve_mol)), add=TRUE)

us_ri_bivalve_pts_latlon <- st_transform(us_ri_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_ri_bivalve_pts_latlon)

farms_us_ri <- data.frame(do.call(rbind, st_geometry(us_ri_bivalve_pts_latlon))) # get lat lon
names(farms_us_ri) <- c("lon", "lat") 
farms_us_ri <- farms_us_ri %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "Rhode Island") 

```

## North Carolina

 - https://deq.nc.gov/about/divisions/marine-fisheries/licenses-permits-and-leases/shellfish-lease-and-franchise#stats-and-figures
 - Comprehensive for shellfish... just need to figure out how download it. 
 - I believe Shellfish is pretty much the only mariculture that occurs in NC. 
 - Has number totals of leases.
 - Download using protocol described for CT.
 - Current as of 12/21/2021
 
North Carolina Division of Marine Fisheries Shellfish Lease and Franchise Program, 2021. North Carolina Division of Marine Fisheries. https://deq.nc.gov/about/divisions/marine-fisheries/licenses-permits-and-leases/shellfish-lease-and-franchise#stats-and-figures

 
```{r}
url <- "https://services2.arcgis.com/kCu40SDxsCGcuUWO/arcgis/rest/services/Lease_DB_View/FeatureServer/2"
nc_shellfish <- esri2sf(url)

nc_active <- nc_shellfish %>%
  filter(Status %in% c("Active", "Approved")) # 423 active leases... wow! 

us_nc_bivalve_mol <- st_transform(nc_active, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_nc_bivalve_pts <- st_centroid(us_nc_bivalve_mol)
# check
plot(st_geometry(us_nc_bivalve_mol))
plot(st_geometry(st_centroid(us_nc_bivalve_mol)), add=TRUE)

us_nc_bivalve_pts_latlon <- st_transform(us_nc_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_nc_bivalve_pts_latlon)

farms_us_nc <- data.frame(do.call(rbind, st_geometry(us_nc_bivalve_pts_latlon))) # get lat lon
names(farms_us_nc) <- c("lon", "lat") 
farms_us_nc <- farms_us_nc %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "North Carolina") 

```
 


## Alaska

- https://adfg.maps.arcgis.com/apps/webappviewer/index.html?id=f3ca95493c1042b39e42a3ecb5dcad6a
 - Will be considered "comprehensive". 
 - Up-to-date as of 2020
 - Specifies species type
 - Csvs with lat lons, but have a "corner" dataset, which we could make into polygons. Has area associated with each site. 
 - Shown in red. 
 - Has 62 active sites. 
 
 
Alaska Department of Fish and Game Aquatic Farming Operations, 2020. Alaska Department of Fish and Game. https://adfg.maps.arcgis.com/apps/webappviewer/index.html?id=f3ca95493c1042b39e42a3ecb5dcad6a


 
```{r}
alaska_details <- read.csv(file.path("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/USA/Alaska/Active Aquatic Farming Operation Details.csv"))

farms_us_ak <- alaska_details %>%
  dplyr::select("lat" = "LATITUDE", "lon" = "LONGITUDE") %>%
  dplyr::mutate(iso3c = "USA", data_quality = 4, subregion = "Alaska")


```
 

## New Hampshire

 - https://granit.unh.edu/data/search
 - Comprehensive
 - Up-to-date as of 2020
 - Polygons 
 - 40 sites
 
NHDES Marine Aquaculture Data, 2020. New Hampshire Department of Environmental Services. https://granit.unh.edu/data


```{r}
nh_aq <- st_read(file.path("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/USA/New_Hampshire/GRANIT_20211011184728/2020_Aquaculture_Sites.shp")) %>% # 40 sites, assume all active
  dplyr::filter(Shellfish == "Yes") # 39 sites

us_nh_bivalve_mol <- st_transform(nh_aq, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_nh_bivalve_pts <- st_centroid(us_nh_bivalve_mol)
# check
plot(st_geometry(us_nh_bivalve_mol))
plot(st_geometry(st_centroid(us_nh_bivalve_mol)), add=TRUE)

us_nh_bivalve_pts_latlon <- st_transform(us_nh_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_nh_bivalve_pts_latlon)

farms_us_nh <- data.frame(do.call(rbind, st_geometry(us_nh_bivalve_pts_latlon))) # get lat lon
names(farms_us_nh) <- c("lon", "lat") 
farms_us_nh <- farms_us_nh %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "New Hampshire") 
```


## Connecticut

Map combined with NY, this is what we need (seaweed and shellfish): https://www.arcgis.com/apps/webappviewer/index.html?id=f8799cefbb4c4751a209710d14b9ad46

Current as of 2019

New York and Connecticut’s Shellfish and Seaweed Aquaculture Viewer, 2019. New York Department of Environmental Conservation. https://www.arcgis.com/apps/webappviewer/index.html?id=f8799cefbb4c4751a209710d14b9ad46 

Roadmap for polygons

Viewer: https://www.arcgis.com/apps/webappviewer/index.html?id=f8799cefbb4c4751a209710d14b9ad46

Item data endpoint (just replace id): https://www.arcgis.com/sharing/rest/content/items/f8799cefbb4c4751a209710d14b9ad46/data

Search for layer you want and grab link ("CT Aquaculture Gear): https://services7.arcgis.com/9fAJJI91yoj2y4Yi/arcgis/rest/services/ConnecticutShellfishOnline2018Draft/FeatureServer/1

Now download using esri2sf R package.

```{r}

url <- "https://services7.arcgis.com/9fAJJI91yoj2y4Yi/arcgis/rest/services/ConnecticutShellfishOnline2018Draft/FeatureServer/1"
ct_shellfish <- esri2sf(url) %>%
  st_as_sf() %>%
  filter(Species != "kelp",
         ActivePerm == "yes")


mapview(ct_shellfish)


us_ct_bivalve_mol <- st_transform(ct_shellfish, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_ct_bivalve_pts <- st_centroid(us_ct_bivalve_mol)
# check
plot(st_geometry(us_ct_bivalve_mol))
plot(st_geometry(st_centroid(us_ct_bivalve_mol)), add=TRUE)

us_ct_bivalve_pts_latlon <- st_transform(us_ct_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_ct_bivalve_pts_latlon)

farms_us_ct <- data.frame(do.call(rbind, st_geometry(us_ct_bivalve_pts_latlon))) # get lat lon
names(farms_us_ct) <- c("lon", "lat") 
farms_us_ct <- farms_us_ct %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "Connecticut") 
```



## Alabama

 - Here is a map with Shellfish leases!: https://aldcnr.maps.arcgis.com/apps/webappviewer/index.html?id=a32dad8dacd249ea86bcb80dd951a424


Alabama Marine Resources Division: Shellfish Aquaculture Siting Tool, 2021. Alabama Marine Resources Division. https://aldcnr.maps.arcgis.com/apps/webappviewer/index.html?id=a32dad8dacd249ea86bcb80dd951a424

Roadmap to download

Viewer: https://aldcnr.maps.arcgis.com/apps/webappviewer/index.html?id=a32dad8dacd249ea86bcb80dd951a424

Item data endpoint (just replace id): https://www.arcgis.com/sharing/rest/content/items/a32dad8dacd249ea86bcb80dd951a424/data

Search for layer you want and grab link ("Leases" is the layer you want): https://services7.arcgis.com/jcVUvxpWnajErgCr/arcgis/rest/services/Shellfish_Riparian_Leases_Revised/FeatureServer/0

Download using esri2sf

```{r}

al_oyster <- esri2sf("https://services7.arcgis.com/jcVUvxpWnajErgCr/arcgis/rest/services/Shellfish_Riparian_Leases_Revised/FeatureServer/0") %>%
  st_as_sf() %>%
  dplyr::filter(Status %in% c("Current", "Current Commercial"))


mapview(al_oyster)

us_al_bivalve_mol <- st_transform(al_oyster, crs = "+init=epsg:3035") # convert to equal area projealion to find centroids. 
us_al_bivalve_pts <- st_centroid(us_al_bivalve_mol)
# check
plot(st_geometry(us_al_bivalve_mol))
plot(st_geometry(st_centroid(us_al_bivalve_mol)), add=TRUE)

us_al_bivalve_pts_latlon <- st_transform(us_al_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_al_bivalve_pts_latlon)

farms_us_al <- data.frame(do.call(rbind, st_geometry(us_al_bivalve_pts_latlon))) # get lat lon
names(farms_us_al) <- c("lon", "lat") 
farms_us_al <- farms_us_al %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 4,
         subregion = "Alabama") 
```


## Washington
 - Aquaculture, Office for Coastal Management (OCM), 2018
 - This dataset has shellfish farms for Washington and Rhode Island, we will use the Washington farms 
 
Source: 
Aquaculture, Office for Coastal Management (OCM), 2018. [Accessed January 17, 2020] https://www.fisheries.noaa.gov/inport/item/53129/

```{r}
us_aqua <- st_read(file.path(path,"USA/US/us_aqua/us_aqua.shp"))

mapview(us_aqua$geometry)

unique(us_aqua$state)

## filter for species !=NA, any that are not bivalves, those that are active or conditionally active (seasonal), and filter out those we already have (NY, ME, CT, NH)

us_aqua_species <- us_aqua %>%
   dplyr::filter(!is.na(species)) %>%
  dplyr::filter(!(species %in% c("finfish", "Atlantic salmon", "Atlantic cod", "Seaweed", "Shrimp", "Coral", "clam seed, oyster seed", "Gelidium spp., Macrocystis pyrifera, M. angustifolia, M. intergrifolia, Pelagophycus spp. Euchema uncinatum (male plants only)", "oyster seed", "In progress", "Not Available", "Hatchery", "softshell", "softshell, butter", "softshell, varnish", "butter"))) %>%
  dplyr::filter(state != "Federal waters") %>%
  dplyr::filter(!(status %in% c(NA, "Pending", "Restricted-Relay", "Not Available", "Conditional Closure", "Prohibited"))) %>%
dplyr::filter(!(state %in% c("Maine", "Connecticut", "New York", "North Carolina", "Alabama", "New Hampshire", "Alaska", "Rhode Island"))) ## filter out the states we already have



unique(us_aqua_species$state)

unique(us_aqua_species$species)

unique(us_aqua_species$status)

mapview(us_aqua_species$geometry)


us_aqua_bivalve_mol <- st_transform(us_aqua_species, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
us_aqua_bivalve_pts <- st_centroid(us_aqua_bivalve_mol)
# check
plot(st_geometry(us_aqua_bivalve_mol))
plot(st_geometry(st_centroid(us_aqua_bivalve_mol)), add=TRUE)

us_aqua_bivalve_pts_latlon <- st_transform(us_aqua_bivalve_pts, crs = "+init=epsg:4326") 

mapview(us_aqua_bivalve_pts_latlon)


farms_us_aqua <- data.frame(do.call(rbind, st_geometry(us_aqua_bivalve_pts_latlon))) 
names(farms_us_aqua) <- c("lon", "lat")
farms_us_aqua <- farms_us_aqua %>%
  mutate(iso3c = "USA", subregion = "Washington") %>%
  mutate(data_quality = 4)
```


## Combine all USA data 

```{r}
farms_usa_total <- rbind(farms_us_aqua, farms_us_ak, farms_us_al, farms_us_ct, farms_us_me, farms_us_nc, farms_us_nh, farms_us_ri)

all_farms_sf <- st_as_sf(farms_usa_total, coords = c("lon", "lat"),
                         crs = crs(gadm_latlon), agr = "constant")

mapview(all_farms_sf)

write_csv(farms_usa_total, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_usa.csv"))


## Now lets make a df for the additional points we will fill in later


usa_additional <- data.frame(iso3c = "USA", type = "bivalve", num_additional_farms = 193 , data_quality = 4, total_farms = 1203)
write.csv(usa_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_usa.csv"), row.names = FALSE)
```


## HELCOM shellfish mariculture points - Sweden, Germany

Source: 
HELCOM HOLAS II Dataset: Shellfish mariculture, 2017. The Baltic Marine Environment Protection Commission. [Accessed October 12, 2020]


Sweden: 

Current data is from HELCOM, but they state that some data is missing for Sweden (likely a very small amount, given the production). There is not much information on how many farms there might be, so we will gapfill Sweden later. Data quality = 2


Germany: 

Need to add 10 points to existing 1 point
 - Current data is from HELCOM, which takes from lease data 
 - Data quality = 4

 - However, the only type farmed in germany is the blue mussel, and this source says there are 11 farms there: 
  - https://onlinelibrary.wiley.com/doi/full/10.1111/raq.12465
  - “The German blue mussel aquaculture takes place at the world heritage and National Park of the Wadden Sea. Therein strict restrictions limit the total number of licenses for harvesting mussel seeds and culture areas (Schlauch 1999). In consequence, the number of enterprises, which are organised in two producer associations, is stable at 11 to ten companies since the 1990s.”

```{r}

## Shellfish points
enc <- readr::guess_encoding(file.path(path, "HELCOM_Baltic/Shellfish_mariculture_points/Shellfish_mariculture_points.shp"), n_max = 1000)[,"encoding"] %>% as.character()

# Read in shapefile
farms_baltic <- st_read(file.path(path, "HELCOM_Baltic/Shellfish_mariculture_points/Shellfish_mariculture_points.shp"), options = "ENCODING=WINDOWS-1252") 

mapview(farms_baltic$geometry)


baltic_pts <- st_centroid(farms_baltic)

bal_pts_latlon <- st_transform(baltic_pts, crs = "+init=epsg:4326") %>%
  dplyr::select(geometry, COUNTRY, NAME)

bal_subregions <- bal_pts_latlon %>%
  dplyr::select(COUNTRY, NAME)

farms_bal <- data.frame(do.call(rbind, st_geometry(bal_pts_latlon)))
names(farms_bal) <- c("lon", "lat")

farms_bal <- cbind(farms_bal, bal_subregions) %>%
  mutate(iso3c = ifelse(COUNTRY == "DE", "DEU", "SWE")) %>%
  dplyr::select(lon, lat, iso3c, subregion = NAME) 



## Sweden

farms_swe <- farms_bal %>%
  dplyr::filter(iso3c == "SWE") %>%
  mutate(data_quality = 2)

write_csv(farms_swe, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_swe.csv"))


## Germany

farms_deu <- farms_bal %>%
  dplyr::filter(iso3c == "DEU") %>%
  mutate(data_quality = 4)

write_csv(farms_deu, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_deu.csv"))


deu_additional <- data.frame(iso3c = "DEU", type = "bivalve", num_additional_farms = 10 , data_quality = 4, total_farms = 11)
write.csv(deu_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_deu.csv"), row.names = FALSE)

```

## HELCOM shellfish mariculture areas - Denmark
This data is comprehensive, data quality = 5

Source:
HELCOM HOLAS II Dataset: Shellfish mariculture, 2017. The Baltic Marine Environment Protection Commission. [Accessed October 12, 2020]

```{r}
## SHELLFISH polygons
# shape file
farms_baltic_area <- st_read(file.path(path, "HELCOM_Baltic/Shellfish_mariculture_areas/Shellfish_mariculture_areas.shp"), options = "ENCODING=WINDOWS-1252") 

mapview(farms_baltic_area$geometry)

baltic_pts_area <- st_centroid(farms_baltic_area)

bal_pts_latlon_area <- st_transform(baltic_pts_area, crs = "+init=epsg:4326") %>%
  dplyr::select(geometry)

farms_bal_area <- data.frame(do.call(rbind, st_geometry(bal_pts_latlon_area)))
names(farms_bal_area) <- c("lon", "lat")

farms_bal_area <- farms_bal_area %>%
  mutate(iso3c = "DNK") %>%
  dplyr::select(lon, lat, iso3c) %>%
  mutate(data_quality = 5)


all_farms_sf <- st_as_sf(farms_bal_area, coords = c("lon", "lat"),
                         crs = crs(gadm_latlon), agr = "constant")

mapview(all_farms_sf)

write.csv(farms_bal_area, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_dnk.csv"), row.names = FALSE)
```
 
## Canada
All of these data are taken from specific governmental sources. They are comprehensive. Data quality = 5.

## British colombia
Shellfish hatcheries: 
https://catalogue.data.gov.bc.ca/dataset/shellfish-hatcheries-coastal-resource-information-management-system-crims/resource/f161583a-1d73-4c92-bf8d-ea67727a89d3

Source: 
Shellfish Hatcheries - Coastal Resource Information Management System (CRIMS), 2019. Ministry of Forests, Lands, Natural Resource Operations and Rural Development. [Accessed November 30, 2020] https://catalogue.data.gov.bc.ca/dataset/shellfish-hatcheries-coastal-resource-information-management-system-crims

```{r}

## read shapefile
bc <- st_read(file.path(path, "Canada/CRIMS_SHELLFISH_HATCHERIES/SHLFISH_HT_point.shp"))

mapview(bc)
bc_pts <- st_centroid(bc)

bc_pts_latlon <- st_transform(bc_pts, crs = "+init=epsg:4326") %>%
  dplyr::select(geometry)

farms_bc <- data.frame(do.call(rbind, st_geometry(bc_pts_latlon)))
names(farms_bc) <- c("lon", "lat")
farms_bc <- farms_bc %>%
  mutate(iso3c="CAN") %>%
  mutate(data_quality = 5)

write_csv(farms_bc, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_bc_can.csv"))

```


## Nova Scotia
Obtained data from here: https://data.novascotia.ca/Fishing-and-Aquaculture/Aquaculture-License-and-Lease-GIS-Database/h57h-p9mm

Source: 
Nova Scotia Marine Aquaculture Leases , 2020. . Nova Scotia Fisheries and Aquaculture. [Accessed February 18, 2020] https://data.novascotia.ca/Fishing-and-Aquaculture/Nova-Scotia-Marine-Aquaculture-Leases/h57h-p9mm

```{r}
# library(jsonlite)
# test <- fromJSON("https://data.novascotia.ca/resource/h57h-p9mm.json")

## read shapefile
ns <- st_read(file.path(path,"Canada/Nova_Scotia_Marine_Aquaculture_Sites/geo_export_fcc6f2ef-82bb-4445-98f0-a82a56cd0d00.shp"))
mapview(ns)

unique(ns$aqua_speci)
                                                                                        
# id bivalves: filter out anything that is not a bivalve
not_bivalve <- c("Atlantic Salmon; Arctic Char; Rainbow Trout", "Rainbow Trout; Atlantic Salmon; Arctic Char; Brook Trout", "Rainbow Trout; Arctic Char; Brook Trout; Atlantic Salmon","Rainbow Trout", "Atlantic Salmon; Rainbow Trout; Atlantic Halibut; Atlantic Cod; Haddock", "Atlantic Salmon", "Atlantic Salmon; Rainbow Trout", "Rainbow Trout; Atlantic Salmon"  )

ns_bivalve <- ns %>%
  filter(!(aqua_speci %in% not_bivalve))

ns_bivalve_mol <- st_transform(ns_bivalve, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
ns_pts <- st_centroid(ns_bivalve_mol)

ns_pts_latlon <- st_transform(ns_pts, crs = "+init=epsg:4326") 

mapview(ns_pts_latlon)

farms_ns <- data.frame(do.call(rbind, st_geometry(ns_pts_latlon))) 
names(farms_ns) <- c("lon", "lat")
farms_ns <- farms_ns %>%
  mutate(iso3c="CAN") %>%
  mutate(data_quality = 5) 

write_csv(farms_ns, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_ns_can.csv"))

```

##  Europe - EMODNET: Ireland, Bulgaria, Italy, Norway, Netherlands, Greece, UK, France

Almost all of this data is comprehensive. The only one that is not is the netherlands. We will note that and gapfill netherlands later on. 

https://www.emodnet-humanactivities.eu/search-results.php?dataname=Shellfish+Production
https://www.emodnet-humanactivities.eu/view-data.php

Source: 
EMODnet Human Activities: Shellfish aquaculture, 2019. [Accessed November 30, 2020] https://www.emodnet-humanactivities.eu/search-results.php?dataname=Shellfish+Production


Netherlands
 - Need to add 76 points to existing 6 (a final total of 82 points)
 - Current data is from EMODnet and is NOT comprehensive. Only 6 farms locations.
  - http://www.fao.org/fishery/countrysector/naso_netherlands/en
  - “The shellfish sub-sector is the older and economically more significant of the two and consists of 50 companies growing blue mussels (Mytilus edulis) and 32 companies growing oysters (Ostrea edulis and Crassostrea gigas).”

```{r}

## read in shapefile
eu <- st_read(file.path(path,"EMODnet_HA_Aquaculture_Shellfish_20191126/EMODnet_HA_Aquaculture_Shellfish_20191126.shp"))
mapview(eu)

unique(eu$COUNTRY)
# [1] "France"         "Ireland"        "Bulgaria"       "United Kingdom" "Italy"          "Norway"        
# [7] "Denmark"        "Netherlands"    "Greece"    

unique(eu$FARM_TYPE)

eu_na <- eu %>%
  dplyr::filter(STATUS == "n.a.")
unique(eu_na$COUNTRY)


eu_active <- eu %>%
  dplyr::filter(STATUS == "Active")

unique(eu_active$COUNTRY)

## we will use both "Active" and "n.a." as the status of the farms. I choose this because all of the denmark points are n.a., and they correspond with the HELCOM data that we used. This indicates to me that the farms denoted as "n.a." are probably actually active. 

eu_active_na <- eu %>%
  dplyr::filter(STATUS %in% c("Active", "n.a."))

unique(eu_active_na$COUNTRY)

# [1] "France"         "Ireland"        "Bulgaria"       "United Kingdom" "Italy"          "Norway"        
# [7] "Denmark"        "Netherlands"    "Greece" 

## Now filter for countries that we want: Ireland, France, UK, Bulgaria, Italy, Norway, Netherlands, and Greece. Filter out denmark, since there is overlap between HELCOM and EMODnet
eu_fao_countries <- eu_active_na %>%
  dplyr::filter(!(COUNTRY %in% c("Denmark")))

unique(eu_fao_countries$COUNTRY)
mapview(eu_fao_countries)


eu_pts <- st_centroid(eu_fao_countries)

eu_pts_latlon <- st_transform(eu_pts, crs = "+init=epsg:4326") %>%
  dplyr::select(COUNTRY, geometry)

eu_countries <- eu_pts_latlon %>%
  dplyr::select(COUNTRY)

farms_eu <- data.frame(do.call(rbind, st_geometry(eu_pts_latlon)))
names(farms_eu) <- c("lon", "lat")

farms_eu_final <- cbind(farms_eu, eu_countries) %>%
  dplyr::select(-geometry) %>%
  mutate(iso3c = case_when(
    COUNTRY == "Ireland" ~ "IRL",
    COUNTRY == "Bulgaria" ~ "BGR", 
    COUNTRY == "Norway" ~ "NOR", 
    COUNTRY == "Netherlands" ~ "NLD",
    COUNTRY == "Italy" ~ "ITA", 
    COUNTRY == "Greece" ~ "GRC",
    COUNTRY == "France" ~ "FRA",
    COUNTRY == "United Kingdom" ~ "GBR"
  )) %>%
  dplyr::select(-COUNTRY)


farms_eu_no_nld <- farms_eu_final %>%
  filter(iso3c != "NLD") %>%
  mutate(data_quality = 5)

write_csv(farms_eu_no_nld, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_GBR_FRA_IRL_BGR_NOR_ITA_GRC.csv"))

farms_nld <- farms_eu_final %>%
  filter(iso3c == "NLD") %>%
  mutate(data_quality = 4)

write_csv(farms_nld, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_NLD.csv"))

nld_additional <- data.frame(iso3c = "NLD", type = "bivalve", num_additional_farms = 76 , data_quality = 4, total_farms = 82)
write.csv(nld_additional, here("marine/bivalve/bivalve_farms/data/subset_known_add_points_4/add_farms_nld.csv"), row.names = FALSE)

test <- farms_eu_final %>%
  group_by(iso3c) %>%
  summarise(n())

```


## UK; Scotland
 - sticking with emodnet for this, since it is comprehensive for the entire country (UK), even if a year behind the scottish gov reporting. 

```{r}
farms_eu_UK <- eu_fao_countries %>%
  filter(COUNTRY == "United Kingdom") # 358.. but only 259 listed as active. Lets compare to the data from scottish gov.  
mapview(farms_eu_UK)

farms_uk <- farms_eu_final %>%
  filter()

## load in data and select active sites, "Seawater" and "Seawater and Freshwater sites", and any sites that contain "Salmon" "Arctic Char" or "Rainbow Trout"

scotland_csv <- as.data.frame(read_csv(file.path(path, "Scotland/Scotland_site_details_2021_raw.csv")))

# Subset to salmonidae farms
sub_scotland <- scotland_csv %>% 
  dplyr::select(ID = `Marine Scotland Site ID`, Easting, Northing, `Producing in Last 3 Years`, `Aquaculture Type`, `Water Type`, Species, subregion = `Local Authority` ) %>% 
  filter(`Producing in Last 3 Years` == c("Yes", "New"),
         `Water Type` %in% c("Seawater", "Freshwater and Seawater"),
         str_detect(Species, "mussel|scallop|Oyster|oyster|urchin|Urchin|Scallop|clam|Clam")) # Only 164 reporting actgive here... however, emodnet reports sites not in scotland and in other places in the uk... so i think we should stick with that actually... especially since they grab their data from the same source.. even if outdated. 

 
```


## Spain

https://servicio.pesca.mapama.es/acuivisor/

 - This is a map of all aquaculture facilities in Spain. It will not allow for download of the locations of the facilities. If we could download tha exact locations, this would be ideal. However, since we can't, we will gapfill locations instead. See archive for how we previously attempted. 
 
Spain bivalves will be treated as if there is no available information, data quality = 1. 


```{r}
# Write a file for russia to indicate it has no information and will need to be fully gapfilled
esp_gf <- data.frame(iso3c = "ESP", type = "bivalve", num_additional_farms = NA, data_quality = 1)
write_csv(esp_gf, here("marine/bivalve/bivalve_farms/data/no_info_1/gf_esp.csv"))
```


 
## Guernsey 
 - https://www.gov.gg/fishing#:~:text=Sea%20Fisheries%20is%20responsible%20for,one%20licensed%20site%20on%20Herm.
 - Sea Fisheries is responsible for the licensing of all shellfish farms within the Bailiwick. There are a total of seven licensed aquaculture sites distributed around Guernsey's west and north coasts, as well as one licensed site on Herm. There is also an oyster farm located in a disused quarry in the north of the island.
 - Allocate 7 farms into the most suitable locations in Guernsey (this is done later, along with the rest of the gapfilling.)
 - For now we will flag this so that we can gapfill later. 
 
Source: 
Fishing and Aquaculture, 2020. Government of the States of Guernsey. [Accessed December 3, 2020] https://www.gov.gg/fishing#:~:text=Sea%20Fisheries%20is%20responsible%20for,one%20licensed%20site%20on%20Herm

```{r}
farms_ggy <- data.frame(iso3c = "GGY", type = "bivalve", num_additional_farms = 7, data_quality = 3)


write_csv(farms_ggy, here("marine/bivalve/bivalve_farms/data/known_number_farms_3/farms_ggy.csv"))

```


## Slovenia
 - http://www.fao.org/fishery/countrysector/naso_slovenia/en#:~:text=From%20the%20345%20aquaculture%20facilities,facilities%20are%20for%20shellfish%20farming.
 - From the 345 aquaculture facilities in Slovenia, there are 26 marine aquaculture facilities, 1 of which is for fish farming, while the remaining **20 facilities are for shellfish farming**. 
 
 
 - Allocate 20 farms into the most suitable locations in Slovenia (this is done later, along with the rest of the gapfilling.)
 - For now we will flag this so that we can gapfill later. 
 
Source: 
National Aquaculture Sector Overview. Slovenia, 2005. , National Aquaculture Sector Overview Fact Sheets. Food and Agriculture Organization of the United Nations.

```{r}
farms_svn <- data.frame(iso3c = "SVN", type = "bivalve", num_additional_farms = 20, data_quality = 3)


write_csv(farms_svn, here("marine/bivalve/bivalve_farms/data/known_number_farms_3/farms_svn.csv"))
```


## Croatia
 - This states there are 266 shellfish farm locations in Croatia: https://ec.europa.eu/fisheries/sites/fisheries/files/docs/pages/spatial-planning-for-aquaculture-hr.pdf
 - However, I can't find much else information. Due to this, I will allocate the 266 farms equally along the coast of Croatia.
 
  - Allocate 266 farms into the most suitable locations in Croatia (this is done later, along with the rest of the gapfilling.)
 
Source: 
Lukin, A., 2014. Spatial Planning For Marine Aquaculture, Croatia, Zadar County Case. Croatian Ministry of Agriculture, Directorate of Fisheries.

```{r}
farms_hrv <- data.frame(iso3c = "HRV", type = "bivalve", num_additional_farms = 266, data_quality = 3)


write_csv(farms_hrv, here("marine/bivalve/bivalve_farms/data/known_number_farms_3/farms_hrv.csv"))
```
 
 
## Cuba
 - Four sites taken from this paper: https://scielo.conicyt.cl/scielo.php?script=sci_arttext&pid=S0718-560X2017000100024
 
Cayo Esquivel del Sur: 23.034188, -80.082047
Cayo de la Cruz: 22.994283, -79.994291
Cayo Hachuela: 22.952677, -79.963228
Cayos de la Enfermeria: 22.90988409545137, -79.94865189587235

Source: 
Mazon Suastegui, J.M., Rivero Suarez, S.E., Betanzos Vega, A., Saucedo, P.E., Rodriguez Jaramillo, C., Acosta Salmon, H., 2017. Potential of sites in northern Cuba for developing an industry of the native mangrove oyster (Crassostrea rhizophorae). lajar 45, 218–222. https://doi.org/10.3856/vol45-issue1-fulltext-24


However, these are just a subset of possible locations in Cuba. Because there is no other information, we will have to gapfill the rest of the possible locations based on a global tonnes/farm. Data quality = 2
 
```{r}

farms_cub <- data.frame(lat = c(23.0341889, 22.994283, 22.952677, 22.90988409545137),
                        lon = c(-80.082047, -79.994291, -79.963228, -79.94865189587235)) %>%
  mutate(iso3c = "CUB") %>%
  mutate(data_quality = 2)

write_csv(farms_cub, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_cub.csv"))

## visualize it 
farms_cub <-  st_as_sf(farms_cub, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_cub)
```

 
## Taiwan
- I can't find any information on the locations of shellfish farms in Taiwan. 
 - https://www.newsecuritybeat.org/2019/10/taiwan-beach-snow/
 
 - Because of this, we will have to gapfill all points based on global tonnes/farm. Data quality = 1
 
Source:
NA
 
 
```{r}

twn_gf <- data.frame(iso3c = "TWN", type = "bivalve", num_additional_farms = NA, data_quality = 1)
write_csv(twn_gf, here("marine/bivalve/bivalve_farms/data/no_info_1/gf_twn.csv"))

```
 
 
## French Polynesia
 - http://www.fao.org/3/t5816e/t5816e04.htm

 - http://www.ressources-marines.gov.pf/aquaculteurs/laquaculture-en-polynesie-francaise/les-filieres-de-laquaculture-en-polynesie-francaise/
 - Aquaculture is: for molluscs: the giant clam (Tridacna maxima).
 - To date (2019), two lagoons are authorized for the collection of giant clam spat, the lagoon of Reao and that of Tatakoto.
 
Backed up by this: https://aquaculture.spc.int/index.php?option=com_countries&view=country&id=6
Giant clams (Tridacna mamoibroodies_bxima) spat collection and stock enhancement techniques are mastered in Eastern Tuamotu atolls. Small scale collectors should be born in 2007.


I will place the two points at these two lagoons (the production is really small, 1200 tonnes):
 - Tatakoto: -17.349601342130857, -138.40992718639146
 - Reao: -18.50754847683744, -136.38467411084235
 
Source: 
The aquaculture sectors in French Polynesia , 2019. Department of Marine Resources Pu Fa’Ahotu Moana. http://www.ressources-marines.gov.pf/aquaculteurs/laquaculture-en-polynesie-francaise/les-filieres-de-laquaculture-en-polynesie-francaise/


 - We are treating this information as comprehensive. 
 
```{r}
farms_pyf <- data.frame(lat = c(-17.349601342130857, -18.50754847683744),
                        lon = c(-138.40992718639146, -136.38467411084235)) %>%
  mutate(iso3c = "PYF") %>%
  mutate(data_quality = 5)

write_csv(farms_pyf, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_pyf.csv"))

## visualize it 
farms_pyf <-  st_as_sf(farms_pyf, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_pyf)
```

 
 
## India 
https://www.researchgate.net/publication/285488252_Mussel_Farming_and_Its_Potential_in_India
 - Has a potential map.
 
http://www.fao.org/fishery/countrysector/naso_india/en
 - Among maritime states, Kerala was the first to recognize the advantages of utilizing mussel farming technology in rural development, from a meagre production in 1997 where cultured mussel production rose to 1 250 tonnes in 2002 with over 250 mussel farms being established in the estuaries of Kerala.
 
 - This is only a subset of existing points... we know there are AT LEAST 250 farms in the state of Kerala, but there could be more elsewhere. Based on this, we will estimate the remaining number of farms for India bivalves based on a global tonnes/farm estimate. 


Source:
FAO Fisheries & Aquaculture - National Aquaculture Sector Overview - India, 2014. National Aquaculture Sector Overview Fact Sheets.

```{r}

farms_ind <- data.frame(iso3c = "IND", type = "bivalve", n_existing_farms = 250, data_quality = 2.5)

write_csv(farms_ind, here("marine/bivalve/bivalve_farms/data/known_subset_2.5/farms_ind.csv"))

```


## Thailand
NASO Maps by FAO: 
http://www.fao.org/fishery/naso-maps/selected-aquaculture-sites/thailand/en/
 - This shows both Shrimp and Bivalve aquaculture sites in Thailand. Although it is not comprehensive, it is the best information I could find. I'll filter for the bivalve sites.
 
 - Because this is not comprehensive, we will have to gapfill the remaining farms based on a global tonnes/farm estimate. This will be done later, but we will flag this as data_quality = 2 for now. 
 
Source:
NASO aquaculture maps collection, 2010. Food and Agriculture Organization of the United Nations. [Accessed January 17, 2020] http://www.fao.org/fishery/naso-maps/selected-aquaculture-sites/thailand/en/

```{r}
## read in thailand shp file
thai_file <- file.path(path, "Thailand/Thailand.kmz")

thai_aqua <- read_sf(thai_file)

mapview(thai_aqua)

## We need to filter for just bivalve aquaculture. The best way that I found to do this was to look at the .kmz file in google earth pro, and determine with rows are bivalves..

## NASO 4,5,6,7 are the bivalves:
unique(thai_aqua$Name)

thai_bivalve <- thai_aqua %>%
  dplyr::filter(Name %in% c("[NASO / 4]", "[NASO / 5]", "[NASO / 6]", "[NASO / 7]"))

mapview(thai_bivalve)

tha_bivalve_mol <- st_transform(thai_bivalve, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
tha_pts <- st_centroid(tha_bivalve_mol)

tha_pts_latlon <- st_transform(tha_pts, crs = "+init=epsg:4326") 

mapview(tha_pts_latlon)

farms_tha <- data.frame(do.call(rbind, st_geometry(tha_pts_latlon))) 
names(farms_tha) <- c("lon", "lat")
farms_tha <- farms_tha %>%
  mutate(iso3c = "THA") %>%
  mutate(data_quality = 2) 


write_csv(farms_tha, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_tha.csv"))


farms_tha <-  st_as_sf(farms_tha, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_tha)
```

 
## Indonesia 
http://www.fao.org/3/AB716E/AB716E08.htm#:~:text=The%20commercially%20important%20bivalve%20species,is%20largely%20experimental%20at%20present.&text=Total%20shellfish%20production%20for%201984,for%20about%202.2%20%25%20of%20this.

 - This has a map of major oyster capture and culture areas in Indonesia... This will have to do. Approximate locations listed below.
 
 From left to right on the map: 
 3.2182751224662542, 99.69952320753342
 -5.964150035506321, 106.35924782378112
 -6.2646012824936275, 108.0659620387556
 -6.75680397048712, 108.64496221771543
 -6.975573928794459, 112.61871244712762
 -4.984732967006041, 120.31326531641196
 0.8890896205268102, 123.52557614041675
 -3.436748178542682, 129.1920950239657
 -5.142358892765042, 137.4727448207172
 
Source:
NETWORK OF AQUACULTURE CENTRES IN ASIA: Status of oyster culture in selected Asian countries, 1988.


  - Because this is not comprehensive, we will have to gapfill the remaining farms based on a global tonnes/farm estimate. This will be done later, but we will flag this as data_quality = 2 for now. 
 
```{r}
farms_idn <- data.frame(lat = c(3.2182751224662542, -5.964150035506321, -6.2646012824936275, -6.75680397048712, -6.975573928794459, -4.984732967006041, 0.889089620526810, -3.436748178542682, -5.142358892765042),
                        lon = c(99.69952320753342, 106.35924782378112, 108.0659620387556, 108.64496221771543, 112.61871244712762, 120.31326531641196, 123.52557614041675, 129.1920950239657, 137.4727448207172)) %>%
  mutate(iso3c = "IDN") %>%
  mutate(data_quality = 2)

write_csv(farms_idn, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_idn.csv"))

## visualize it 
farms_idn <-  st_as_sf(farms_idn, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_idn)


```


## Russia
 - This probably doesn't really matter, because the production is so low, and the coastline is so big...
 - I can't find any good information on shellfish farming in Russia. I will place a point along the coast.
 - Based on this, we will have to gapfill based on a global tonnes/farm estimate to get a proper number of farm points. Then we will place those points in suitable locations for aquaculture. 
 
Source:
NA

```{r}
rus_gf <- data.frame(iso3c = "RUS", type = "bivalve", num_additional_farms = NA, data_quality = 1)
write_csv(rus_gf, here("marine/bivalve/bivalve_farms/data/no_info_1/gf_rus.csv"))

```
 

## Korea, Dem. People's Rep
**Revisiting Changing Patterns of North Koreaʹs Fisheries Production: 1990s‐2000s**
"By the 1980s, there has been 60 or more fish farming companies in North
Korea's national distribution" 

 - Because this is a very outdated source, and likely includes some fish farms, we will include the 60 fish farms, and later gapfill based on a global tonnes/farm, like we are going to do for India.  
 
 
Source:
Park, S., Hong, S., 2012. Revisiting Changing Patterns of North Koreaʹs Fisheries Production: 1990s‐2000s 19.

```{r}
farms_prk <- data.frame(iso3c = "PRK", type = "bivalve", n_existing_farms = 60, data_quality = 2.5)

write_csv(farms_prk, here("marine/bivalve/bivalve_farms/data/known_subset_2.5/farms_prk.csv"))
```
 
 
## Brazil 
http://www.fao.org/fishery/countrysector/naso_brazil/en
"Mollusc farming was introduced into Brazil in the 1960s but only since 1989 has this activity developed as an important economic alternative for small-scale fishermen. Besides the generation of income and employment, bivalve culture has had a noticeable effect in contributing to the stabilisation of traditional communities on their original lands. Today, there are no large companies in this sector and small farmers are organized into associations and cooperatives. Santa Catarina State in the South region is responsible for 93 percent of Brazilian bivalve mollusc production."
 
https://onlinelibrary.wiley.com/doi/full/10.1111/raq.12107
 "Currently, the total bivalve production reaches 19 082 tons, with **729 marine farmers** scattered among the 12 municipalities mentioned above, and marine aquaculture has now reached an important position in Santa Catarina State as a significant contributor to food production, providing employment and livelihood opportunities and representing a legitimate use of the marine environment for sustainable food production"
 
 
Of the 837 aquaculture areas grouped inside the **27 aquaculture parks** originally proposed by the PLDM, 798 areas and 24 parks are already licensed and attended by **618 shellfish farmers** who represent 74% of all shellfish farmers in Santa Catarina State 


 - Based on this paper, I will assign 27 points along the coastline of Santa Catarina state. I'm assuming that the 3 remaining unlicensed aquaculture parks are licensed now. 
 
Source:
Suplicy, F.M., Vianna, L.F. de N., Rupp, G.S., Novaes, A.L.T., Garbossa, L.H.P., de Souza, R.V., Guzenski, J., da Costa, S.W., Silva, F.M., dos Santos, A.A., 2017. Planning and management for sustainable coastal aquaculture development in Santa Catarina State, south Brazil. Rev Aquacult 9, 107–124. https://doi.org/10.1111/raq.12107

 - Although outdated, I will treat this as truth. The FAO backs up that the majority of shell fish farming occurs in Santa Catarina state. Data_quality = 3 
 
```{r}

farms_bra <- data.frame(iso3c = "BRA", type = "bivalve", num_additional_farms = 27, data_quality = 3)


write_csv(farms_bra, here("marine/bivalve/bivalve_farms/data/known_number_farms_3/farms_bra.csv"))
```
 
 
## Philippines

http://www.fao.org/3/AB710E/AB710E18.htm#ann12
 - This has the number of farms per province. Although a little dated, we will allocate this number (556) of farms into the suitable locations

LOCATION	Number of farms	
Zambales	 	 	NA
Bulacan	 	 	NA
Mindoro Oriental	 NA
Palawan	 	 	NA
Quezon	 	  NA
Capiz	**300**		NA
Iloilo	 	 	NA
Aklan	 	 	NA
Negros Occidental	**6** 
Cebu	 	 	NA
Bohol	 	 	NA
Leyte	 	 	NA
Samar	**250**	
Total	**556**

Source:
REPORT OF THE WORKSHOP AND STUDY TOUR ON MOLLUSC SANITATION AND MARKETING, 1989. . Food and Agriculture Organization of the United Nations.



```{r}
farms_phl <- data.frame(iso3c = "PHL", type = "bivalve", num_additional_farms = 556, data_quality = 3)


write_csv(farms_phl, here("marine/bivalve/bivalve_farms/data/known_number_farms_3/farms_phl.csv"))
```


## Malaysia
http://www.fao.org/fishery/countrysector/naso_malaysia/en
"Brackish water aquaculture is characterized by the extensive culture of bivalve molluscs, occupying an area of 7 659 ha, mostly in the western coastal waters where there is an abundance of mud flats suitable for the culture of blood cockles"

 - Because of a lack of information, this country will need to be gapfilled based on a global tonnes/farm estimate to get an estimate number of farm points to place in suitable locations. 
 
Source: 
NA

```{r}
mys_gf <- data.frame(iso3c = "MYS", type = "bivalve", num_additional_farms = NA, data_quality = 1)
write_csv(mys_gf, here("marine/bivalve/bivalve_farms/data/no_info_1/gf_mys.csv"))


```

## Cambodia 
http://www.fao.org/fishery/countrysector/naso_cambodia/en
 - Because of a lack of information, this country will need to be gapfilled based on a global tonnes/farm estimate to get an estimate number of farm points to place in suitable locations. 

 
Source:
NA

```{r}

khm_gf <- data.frame(iso3c = "KHM", type = "bivalve", num_additional_farms = NA, data_quality = 1)
write_csv(khm_gf, here("marine/bivalve/bivalve_farms/data/no_info_1/gf_khm.csv"))

```


## Hong Kong
https://www.afcd.gov.hk/english/fisheries/fish_aqu/fish_aqu_mpo/fish_aqu_mpo.html

All farms are located in Deep Bay, specifically shown in the map. 

 - I will place a point in Deep Bay. Data quality = 5
 
Source: 
Marine fish culture, pond fish culture and oyster culture, n.d. Agriculture, Fisheries, and Conservation Department; The Government of the Hong Kong Special Administrative Region.  https://www.afcd.gov.hk/english/fisheries/fish_aqu/fish_aqu_mpo/fish_aqu_mpo.html 
 
```{r}


farms_hkg <- data.frame(lat = c(22.48384043911231),
                        lon = c(113.98143449612208)) %>%
  mutate(iso3c = "HKG") %>%
  mutate(data_quality = 5)

write_csv(farms_hkg, here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5/farms_hkg.csv"))

## visualize it 
farms_hkg <-  st_as_sf(farms_hkg, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_hkg)

```

 
# Join cell count dataframes

```{r}
farm_files <- list.files(here("marine/bivalve/bivalve_farms/data/known_farm_sites_2_4_5"), full=TRUE)
all_farms = plyr::ldply(farm_files, read_csv)

all_farms_1 <- all_farms %>%
  dplyr::select(-data_quality)

write.csv(all_farms_1, here("marine/bivalve/bivalve_farms/data/global_bivalve_farm_lat_lon.csv"), row.names = FALSE)

write.csv(all_farms, here("marine/bivalve/bivalve_farms/data/global_bivalve_farm_lat_lon_quality.csv"), row.names = FALSE)

# all_farms <- read_csv(here("marine/salmon/salmon_farm/data/global_salmon_farm_lat_lon_quality.csv"), col_types = "ddccd")

all_farms_sf <- st_as_sf(all_farms, coords = c("lon", "lat"), 
                         crs = crs(gadm_latlon), agr = "constant")

mapview(all_farms_sf)

unique(all_farms$iso3c)
```
